discovery/dns: move dns to own package
